refact: Add scraper_utils.py
Add module for general utility functions used for scraping needs.
